continue-playback: Move tracker.py to internal
